+++ src/cpc/cpc/ByteFrameCtrl.ts
@@ -1,27 +1,43 @@
 import { trans } from "../core/mod.ts";
 import { RpcFrame } from "../core/type.ts";
-import { RpcFrameCtrl } from "./cpc.ts";
+import { CpcController, CpcFrameSource } from "./cpc.ts";
 
 /**
- * 收集发送的帧，在下一个微任务触发发送
+ * 自动处理二进制帧。它会将 CpCall 在一个宏任务内的生成的帧进行打包
  */
-export class ByteFrameCtrl implements RpcFrameCtrl {
-  constructor(private ctrl: RpcFrameCtrl<Uint8Array>) {
-    this.frameIter = trans.createFrameIterator(ctrl.frameIter);
+export class CpcByteFrameSource implements CpcFrameSource<RpcFrame> {
+  constructor(private ctrl: CpcFrameSource<Uint8Array>) {}
+  init({ endFrame, nextFrame }: CpcController): void {
+    const parser = trans.createCpcFrameParser();
+    this.ctrl.init({
+      endFrame: endFrame,
+      nextFrame: function (buf): boolean {
+        while (parser.next(buf)) {
+          const res = parser.finish();
+          const frame = trans.decodeCpcFrame(res.value).frame;
+          nextFrame(frame);
+          if (res.residue) buf = res.residue;
+          else break;
+        }
+        return false;
+      },
+    });
   }
   close(): void | Promise<void> {
-    this.ctrl.close?.();
+    if (this.link.length) this.send();
+    return this.ctrl.close();
   }
   private link: RpcFrame[] = [];
   dispose(reason?: any): void {
     this.ctrl.dispose?.(reason);
   }
-  send() {
-    const chunk = trans.packCpcFrames(this.link);
-    this.ctrl.sendFrame(chunk);
+  private send() {
+    if (this.link.length) {
+      const chunk = trans.packCpcFrames(this.link);
+      this.ctrl.sendFrame(chunk);
+    }
     this.link.length = 0;
   }
-  frameIter: AsyncIterable<RpcFrame>;
   sendFrame(frame: RpcFrame): void {
     if (this.link.length === 0) Promise.resolve().then(() => this.send());
     this.link.push(frame);